Curl_getconnectinfo: fixed to work with always-multi
To make sure CURLINFO_LASTSOCKET works and curl_easy_recv() etc when the
connection cache is now always in the multi handle that the easy handle
doesn't have a strong connection to anymore, the code now searches for
the last used connection in the last used multi handle to see if it is
still there and then gets the socket from there.

This makes CURLINFO_LASTSOCKET only work if curl_easy_perform() has been
used on an easy handle, but that was always how the feature was
documented.

As a result of this restructure, I've created multihandle.h which now
has the multi handle struct and related stuff defined so that more files
than multi.c can use it.